Wrote csv->json function
Wrote a function that read the csv file and extract a single feature of a single exon. The output data is formatted for hierarchical bar chart and written to a json file. The next step is to wrote another funcion that iterate through the csv file and extract all the features of one exon. ps. I am using the deciphering_rna_splicing_single_exon.csv file.