Completed csv -> json function
Based on the previous single column reader, iterate through other incl and skip columns, format, and output as a tree structure. The function now takes the file path of the csv file of an exon and will be able to produce the tree structure that is essential for building the hierarchical bar chart later on.